Datasets

Datasets

Unlabeled Datasets

Labeled Datasets

Machine learning datasets are a collection of instances (or samples), each one described by a number of variables. In the case of tabular data, a dataset looks like a database table, where every column is a variable, and each row corresponds to a given instance. However, a dataset can also be non-tabular (or unstructured); for example, each instance can consist of a multivariate time-series, or an image.

When data is composed of different modalities) combining their statistical properties is non-trivial, since they may be quite different in nature one another.

The abstract representation of a multimodal dataset provided by this package is the AbstractMultiModalDataset.

SoleData.AbstractMultiModalDataset — Type

Abstract supertype for all multimodal datasets.

A concrete multimodal dataset should always provide accessors data, to access the underlying tabular structure (e.g., DataFrame) and grouped_variables, to access the grouping of variables (a vector of vectors of column indices).

SoleData.grouped_variables — Function

grouped_variables(amd)::Vector{Vector{Int}}

Return the indices of the variables grouped by modality, of an AbstractMultiModalDataset. The grouping describes how the different modalities are composed from the underlying AbstractDataFrame structure.

See also data, AbstractMultiModalDataset.

SoleData.data — Function

data(amd)::AbstractDataFrame

Return the structure that underlies an AbstractMultiModalDataset.

See also grouped_variables, AbstractMultiModalDataset.

SoleData.dimension — Function

dimension(df)

Return the dimension of a dataframe df.

If the dataframe has variables of various dimensions :mixed is returned.

If the dataframe is empty (no instances) :empty is returned. This behavior can be changed by setting the keyword argument force:

:no (default): return :mixed in case of mixed dimension

:max: return the greatest dimension

:min: return the lowest dimension

Unlabeled Datasets

In unlabeled datasets there is no label variable, and all of the variables (also called feature variables, or features) have equal role in the representation. These datasets are used in unsupervised learning contexts, for discovering internal correlation patterns between the features. Multimodal unlabeled datasets can be instantiated with MultiModalDataset.

SoleData.MultiModalDataset — Type

MultiModalDataset(grouped_variables, df)

Create a MultiModalDataset from a DataFrame, df, initializing modalities accordingly to grouped_variables parameter.

grouped_variables is an AbstractVector of modality descriptor which are AbstractVectors of integers representing the index of the variables selected for that modality.

The order matters for both the indices and the variables in them.

julia> df = DataFrame(
                  :age => [30, 9],
                  :name => ["Python", "Julia"],
                  :stat1 => [[sin(i) for i in 1:50000], [cos(i) for i in 1:50000]],
                  :stat2 => [[cos(i) for i in 1:50000], [sin(i) for i in 1:50000]]
              )
2×4 DataFrame
 Row │ age    name    stat1                              stat2                             ⋯
     │ Int64  String  Array…                             Array…                            ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────
   1 │    30  Python  [0.841471, 0.909297, 0.14112, -0…  [0.540302, -0.416147, -0.989992,… ⋯
   2 │     9  Julia   [0.540302, -0.416147, -0.989992,…  [0.841471, 0.909297, 0.14112, -0…

julia> md = MultiModalDataset([[2]], df)
● MultiModalDataset
   └─ dimensions: (0,)
- Modality 1 / 1
   └─ dimension: 0
2×1 SubDataFrame
 Row │ name
     │ String
─────┼────────
   1 │ Python
   2 │ Julia
- Spare variables
   └─ dimension: mixed
2×3 SubDataFrame
 Row │ age    stat1                              stat2
     │ Int64  Array…                             Array…
─────┼─────────────────────────────────────────────────────────────────────────────
   1 │    30  [0.841471, 0.909297, 0.14112, -0…  [0.540302, -0.416147, -0.989992,…
   2 │     9  [0.540302, -0.416147, -0.989992,…  [0.841471, 0.909297, 0.14112, -0…

MultiModalDataset(df; group = :none)

Create a MultiModalDataset from a DataFrame, df, automatically selecting modalities.

Selection of modalities can be controlled by the parameter group which can be:

:none (default): no modality will be created

:all: all variables will be grouped by their dimension

a list of dimensions which will be grouped.

Note: :all and :none are the only Symbols accepted by group.

TODO: :all should be default parameter value for group

TODO: fix passing a vector of Integer to group parameter

TODO: rewrite examples

EXAMPLES

julia> df = DataFrame(
                  :age => [30, 9],
                  :name => ["Python", "Julia"],
                  :stat1 => [[sin(i) for i in 1:50000], [cos(i) for i in 1:50000]],
                  :stat2 => [[cos(i) for i in 1:50000], [sin(i) for i in 1:50000]]
              )
2×4 DataFrame
 Row │ age    name    stat1                              stat2                             ⋯
     │ Int64  String  Array…                             Array…                            ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────
   1 │    30  Python  [0.841471, 0.909297, 0.14112, -0…  [0.540302, -0.416147, -0.989992,… ⋯
   2 │     9  Julia   [0.540302, -0.416147, -0.989992,…  [0.841471, 0.909297, 0.14112, -0…

julia> md = MultiModalDataset(df)
● MultiModalDataset
   └─ dimensions: ()
- Spare variables
   └─ dimension: mixed
2×4 SubDataFrame
 Row │ age    name    stat1                              stat2                             ⋯
     │ Int64  String  Array…                             Array…                            ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────
   1 │    30  Python  [0.841471, 0.909297, 0.14112, -0…  [0.540302, -0.416147, -0.989992,… ⋯
   2 │     9  Julia   [0.540302, -0.416147, -0.989992,…  [0.841471, 0.909297, 0.14112, -0…

julia> md = MultiModalDataset(df; group = :all)
● MultiModalDataset
   └─ dimensions: (0, 1)
- Modality 1 / 2
   └─ dimension: 0
2×2 SubDataFrame
 Row │ age    name
     │ Int64  String
─────┼───────────────
   1 │    30  Python
   2 │     9  Julia
- Modality 2 / 2
   └─ dimension: 1
2×2 SubDataFrame
 Row │ stat1                              stat2
     │ Array…                             Array…
─────┼──────────────────────────────────────────────────────────────────────
   1 │ [0.841471, 0.909297, 0.14112, -0…  [0.540302, -0.416147, -0.989992,…
   2 │ [0.540302, -0.416147, -0.989992,…  [0.841471, 0.909297, 0.14112, -0…

julia> md = MultiModalDataset(df; group = [0])
● MultiModalDataset
   └─ dimensions: (0, 1, 1)
- Modality 1 / 3
   └─ dimension: 0
2×2 SubDataFrame
 Row │ age    name
     │ Int64  String
─────┼───────────────
   1 │    30  Python
   2 │     9  Julia
- Modality 2 / 3
   └─ dimension: 1
2×1 SubDataFrame
 Row │ stat1
     │ Array…
─────┼───────────────────────────────────
   1 │ [0.841471, 0.909297, 0.14112, -0…
   2 │ [0.540302, -0.416147, -0.989992,…
- Modality 3 / 3
   └─ dimension: 1
2×1 SubDataFrame
 Row │ stat2
     │ Array…
─────┼───────────────────────────────────
   1 │ [0.540302, -0.416147, -0.989992,…
   2 │ [0.841471, 0.909297, 0.14112, -0…

SoleData._empty — Method

_empty(md)

Return a copy of a multimodal dataset with no instances.

Note: since the returned AbstractMultiModalDataset will be empty its columns types will be Any.

Labeled Datasets

In labeled datasets, one or more variables are considered to have special semantics with respect to the other variables; each of these labeling variables (or target variables) can be thought as assigning a label to each instance, which is typically a categorical value (classification label) or a numerical value (regression label). Supervised learning methods can be applied on these datasets for modeling the target variables as a function of the feature variables.

As an extension of the AbstractMultiModalDataset, AbstractLabeledMultiModalDataset has an interface that can be implemented to represent multimodal labeled datasets.

SoleData.AbstractLabeledMultiModalDataset — Type

Abstract supertype for all labelled multimodal datasets (used in supervised learning).

As any multimodal dataset, any concrete labeled multimodal dataset should always provide the accessors data, to access the underlying tabular structure (e.g., DataFrame) and grouped_variables, to access the grouping of variables. In addition to these, implementations are required for labeling_variables, to access the indices of the labeling variables.

See also AbstractMultiModalDataset.

SoleData.labeling_variables — Function

labeling_variables(almd)::Vector{Int}

Return the indices of the labelling variables, of the AbstractLabeledMultiModalDataset. with respect to the underlying AbstractDataFrame structure (see data).

See also grouped_variables, AbstractLabeledMultiModalDataset.

Missing docstring.

Missing docstring for dataset. Check Documenter's build log for details.

Multimodal labeled datasets can be instantiated with LabeledMultiModalDataset.

SoleData.LabeledMultiModalDataset — Type

LabeledMultiModalDataset(md, labeling_variables)

Create a LabeledMultiModalDataset by associating a AbstractMultiModalDataset with some labeling variables; the labeling_variables are specified as a column index (Int) or a vector of column indices (Vector{Int}).

PARAMETERS

labeling_variables is an AbstractVector of Integers indicating the indices of the variables that will be set as labels;

md is the original AbstractMultiModalDataset.

EXAMPLES

julia> lmd = LabeledMultiModalDataset(MultiModalDataset([[2],[4]], DataFrame(
           :id => [1, 2],
           :age => [30, 9],
           :name => ["Python", "Julia"],
           :stat => [[sin(i) for i in 1:50000], [cos(i) for i in 1:50000]]
       )), [1, 3])
● LabeledMultiModalDataset
   ├─ labels
   │   ├─ id: Set([2, 1])
   │   └─ name: Set(["Julia", "Python"])
   └─ dimensions: (0, 1)
- Modality 1 / 2
   └─ dimension: 0
2×1 SubDataFrame
 Row │ age
     │ Int64
─────┼───────
   1 │    30
   2 │     9
- Modality 2 / 2
   └─ dimension: 1
2×1 SubDataFrame
 Row │ stat
     │ Array…
─────┼───────────────────────────────────
   1 │ [0.841471, 0.909297, 0.14112, -0…
   2 │ [0.540302, -0.416147, -0.989992,…

SoleData.joinlabels! — Method

joinlabels!(lmd, [lbls...]; delim = "_")

With a labeled multimodal dataset, collapse the label variables identified by lbls into a single label variable of type String, by means of a join that uses delim for string delimiter.

If not specified differently this function will join all labels.

lbls can be of type Integer (indicating the index of the label) or the name of the label.

!!! note

The resulting label will always be of type String.

Note

The resulting label will always be added as last column in the underlying DataFrame.

EXAMPLES

julia> lmd = LabeledMultiModalDataset(
           MultiModalDataset(
               [[2],[4]],
               DataFrame(
                   :id => [1, 2],
                   :age => [30, 9],
                   :name => ["Python", "Julia"],
                   :stat => [[sin(i) for i in 1:50000], [cos(i) for i in 1:50000]]
               )
           ),
           [1, 3],
       )
● LabeledMultiModalDataset
   ├─ labels
   │   ├─ id: Set([2, 1])
   │   └─ name: Set(["Julia", "Python"])
   └─ dimensions: (0, 1)
- Modality 1 / 2
   └─ dimension: 0
2×1 SubDataFrame
 Row │ age
     │ Int64
─────┼───────
   1 │    30
   2 │     9
- Modality 2 / 2
   └─ dimension: 1
2×1 SubDataFrame
 Row │ stat
     │ Array…
─────┼───────────────────────────────────
   1 │ [0.841471, 0.909297, 0.14112, -0…
   2 │ [0.540302, -0.416147, -0.989992,…

julia> joinlabels!(lmd)
● LabeledMultiModalDataset
   ├─ labels
   │   └─ id_name: Set(["1_Python", "2_Julia"])
   └─ dimensions: (0, 1)
- Modality 1 / 2
   └─ dimension: 0
2×1 SubDataFrame
 Row │ age
     │ Int64
─────┼───────
   1 │    30
   2 │     9
- Modality 2 / 2
   └─ dimension: 1
2×1 SubDataFrame
 Row │ stat
     │ Array…
─────┼───────────────────────────────────
   1 │ [0.841471, 0.909297, 0.14112, -0…
   2 │ [0.540302, -0.416147, -0.989992,…

SoleData.label — Method

label(lmd, instance, i)

Return the value of the i-th labeling variable for instance at index instance in a labeled multimodal dataset.

SoleData.labeldomain — Method

labeldomain(lmd, i)

Return the domain of i-th label of a labeled multimodal dataset.

SoleData.labels — Method

labels(lmd, instance)
labels(lmd)

Return the labels of instance at index instance in a labeled multimodal dataset. A dictionary of type labelname => value is returned.

If only the first argument is passed then the names of all labels are returned.

SoleData.nlabelingvariables — Method

nlabelingvariables(lmd)

Return the number of labels of a labeled multimodal dataset.

SoleData.removefromlabels! — Method

removefromlabels!(lmd, i)
removefromlabels!(lmd, var_name)

Remove i-th labeling variable from labels list.

The variable name can be passed as second argument instead of its index.

SoleData.setaslabelinging! — Method

setaslabelinging!(lmd, i)
setaslabelinging!(lmd, var_name)

Set i-th variable as label.

The variable name can be passed as second argument instead of its index.