SoleData

Welcome to the documentation for SoleData.

SoleData.BASE_FEATURE_ALIASES

SoleLogics.IA_L

SoleData.AbstractCondition

SoleData.AbstractConditionalAlphabet

SoleData.AbstractFeature

SoleData.AbstractFullMemoset

SoleData.AbstractMemoset

SoleData.AbstractModalLogiset

SoleData.AbstractOneStepMemoset

SoleData.AbstractScalarOneStepRelationalMemoset

SoleData.AbstractUnivariateFeature

SoleData.BoundedScalarConditions

SoleData.ExplicitFeature

SoleData.Feature

SoleData.FullMemoset

SoleData.FunctionalCondition

SoleData.MultiFormula

SoleData.MultiLogiset

SoleData.MultivariateFeature

SoleData.ScalarChainedMemoset

SoleData.ScalarCondition

SoleData.ScalarMetaCondition

SoleData.ScalarOneStepMemoset

SoleData.SupportedLogiset

SoleData.UnboundedScalarConditions

SoleData.UnivariateFeature

SoleData.UnivariateMax

SoleData.UnivariateMin

SoleData.UnivariateNamedFeature

SoleData.UnivariateSoftMax

SoleData.UnivariateSoftMin

SoleData.UnivariateValue

SoleData.ValueCondition

SoleData.VarFeature

SoleLogics.AbstractFrame

SoleLogics.AbstractWorld

SoleLogics.Atom

SoleLogics.Interval

SoleLogics.Interval2D

SoleData.computefeature

SoleData.computeunivariatefeature

SoleData.featvaltype

SoleData.minify

SoleData.parsecondition

SoleData.parsefeature

SoleData.representatives

SoleData.variable_name

SoleLogics.accessibles

SoleLogics.syntaxstring

Logical foundations

See SoleLogics for more.

SoleLogics.Atom — Type

struct Atom{V} <: SyntaxLeaf
    value::V
end

An atom, sometimes called an atomic proposition, propositional letter (or simply letter), of type Atom{V} wraps a value::V representing a fact which truth can be assessed on a logical interpretation.

Atoms are nullary tokens (i.e, they are at the leaves of a syntax tree); note that their atoms cannot be Atoms.

See also AbstractInterpretation, atoms, check, SyntaxToken.

SoleLogics.AbstractWorld — Type

abstract type AbstractWorld end

Abstract type for the nodes of an annotated accessibility graph (Kripke structure). This is used, for example, in modal logic, where the truth of formulas is relativized to worlds, that is, nodes of a graph.

Implementing

When implementing a new world type, the logical semantics should be defined via accessibles methods; refer to the help for accessibles.

See also AbstractKripkeStructure, AbstractFrame.

SoleLogics.Interval — Type

struct Interval{T} <: GeometricalWorld
    x :: T
    y :: T
end

An interval in a 1-dimensional space, with coordinates of type T.

Examples

julia> SoleLogics.goeswithdim(SoleLogics.Interval(1,2),1)
true

julia> SoleLogics.goeswithdim(SoleLogics.Interval(1,2),2)
false

julia> collect(accessibles(SoleLogics.FullDimensionalFrame(5), Interval(1,2), SoleLogics.IA_L))
6-element Vector{Interval{Int64}}:
 (3−4)
 (3−5)
 (4−5)
 (3−6)
 (4−6)
 (5−6)

See also goeswithdim, accessibles, FullDimensionalFrame, Point, Interval2D, GeometricalWorld, AbstractWorld.

SoleLogics.Interval2D — Type

struct Interval2D{T} <: GeometricalWorld
    x :: Interval{T}
    y :: Interval{T}
end

A orthogonal rectangle in a 2-dimensional space, with coordinates of type T. This is the 2-dimensional Interval counterpart, that is, the combination of two orthogonal Intervals.

Examples

julia> SoleLogics.goeswithdim(SoleLogics.Interval2D((1,2),(3,4)),1)
false

julia> SoleLogics.goeswithdim(SoleLogics.Interval2D((1,2),(3,4)),2)
true

julia> collect(accessibles(SoleLogics.FullDimensionalFrame(5,5), Interval2D((2,3),(2,4)), SoleLogics.IA_LL))
3-element Vector{Interval2D{Int64}}:
 ((4−5)×(5−6))
 ((4−6)×(5−6))
 ((5−6)×(5−6))

See also goeswithdim, accessibles, FullDimensionalFrame, Point, Interval, GeometricalWorld, AbstractWorld.

SoleLogics.syntaxstring — Function

syntaxstring(s::Syntactical; kwargs...)::String

Return the string representation of any syntactic object (e.g., Formula, SyntaxTree, SyntaxToken, Atom, Truth, etc). Note that this representation may introduce redundant parentheses. kwargs can be used to specify how to display syntax tokens/trees under some specific conditions.

The following kwargs are currently supported:

function_notation = false::Bool: when set to true, it forces the use of function notation for binary operators (see here).

remove_redundant_parentheses = true::Bool: when set to false, it prints a syntaxstring where each syntactical element is wrapped in parentheses.

parenthesize_atoms = !remove_redundant_parentheses::Bool: when set to true, it forces the atoms (which are the leaves of a formula's tree structure) to be wrapped in parentheses.

Examples

julia> syntaxstring(parseformula("p∧q∧r∧s∧t"))
"p ∧ q ∧ r ∧ s ∧ t"

julia> syntaxstring(parseformula("p∧q∧r∧s∧t"), function_notation=true)
"∧(∧(∧(∧(p, q), r), s), t)"

julia> syntaxstring(parseformula("p∧q∧r∧s∧t"), remove_redundant_parentheses=false)
"((((p) ∧ (q)) ∧ (r)) ∧ (s)) ∧ (t)"

julia> syntaxstring(parseformula("p∧q∧r∧s∧t"), remove_redundant_parentheses=true, parenthesize_atoms=true)
"(p) ∧ (q) ∧ (r) ∧ (s) ∧ (t)"

julia> syntaxstring(parseformula("◊((p∧s)→q)"))
"◊((p ∧ s) → q)"

julia> syntaxstring(parseformula("◊((p∧s)→q)"); function_notation = true)
"◊(→(∧(p, s), q))"

See also parseformula, SyntaxBranch, SyntaxToken.

Implementation

In the case of a syntax tree, syntaxstring is a recursive function that calls itself on the syntax children of each node. For a correct functioning, the syntaxstring must be defined (including the kwargs... part!) for every newly defined SyntaxToken (e.g., SyntaxLeafs, that is, Atoms and Truth values, and Operators), in a way that it produces a unique string representation, since Base.hash and Base.isequal, at least for SyntaxTrees, rely on it.

In particular, for the case of Atoms, the function calls itself on the wrapped value:

syntaxstring(a::Atom; kwargs...) = syntaxstring(value(a); kwargs...)

The syntaxstring for any value defaults to its string representation, but it can be defined by defining the appropriate syntaxstring method.

Warning

The syntaxstring for syntax tokens (e.g., atoms, operators) should not be prefixed/suffixed by whitespaces, as this may cause ambiguities upon parsing. For similar reasons, syntaxstrings should not contain parentheses ('(', ')'), and, when parsing in function notation, commas (',').

See also SyntaxLeaf, Operator, parseformula.

SoleLogics.IA_L — Constant

See IntervalRelation

SoleLogics.AbstractFrame — Type

abstract type AbstractFrame{W<:AbstractWorld} end

Abstract type for an accessibility graph (Kripke frame), that gives the topology to Kripke structures. A frame can be queried for its set of vertices (also called worlds, see allworlds), and it can be browsed via its accessibility relation(s) (see accessibles). Refer to FullDimensionalFrame as an example.

See also truthtype, , allworlds, nworlds, AbstractKripkeStructure, AbstractWorld.

SoleLogics.accessibles — Function

accessibles(fr::AbstractUniModalFrame{W}, w::W)::Worlds{W} where {W<:AbstractWorld}

Return the worlds in frame fr that are accessible from world w.

See also AbstractWorld, AbstractUniModalFrame.

accessibles(
    fr::AbstractMultiModalFrame{W},
    w::W,
    r::AbstractRelation
) where {W<:AbstractWorld}

Return the worlds in frame fr that are accessible from world w via relation r.

Examples

julia> fr = SoleLogics.FullDimensionalFrame((10,),);

julia> typeof(accessibles(fr, Interval(2,5), IA_L))
Base.Generator{...}

julia> typeof(accessibles(fr, globalrel))
Base.Generator{...}

julia> @assert SoleLogics.nworlds(fr) == length(collect(accessibles(fr, globalrel)))

julia> typeof(accessibles(fr, Interval(2,5), identityrel))
Vector{Interval{Int64}}

julia> Interval(8,11) in collect(accessibles(fr, Interval(2,5), IA_L))
true

Implementation

Since accessibles always returns an iterator of worlds of the same type W, the current implementation of accessibles for multi-modal frames delegates the enumeration to a lower level _accessibles function, which returns an iterator of parameter tuples that are, then, fed to the world constructor the using IterTools generators, as in:

function accessibles(
    fr::AbstractMultiModalFrame{W},
    w::W,
    r::AbstractRelation,
) where {W<:AbstractWorld}
    IterTools.imap(W, _accessibles(fr, w, r))
end

As such, when defining new frames, worlds, and/or relations, one should provide new methods for _accessibles. For example:

_accessibles(fr::Full1DFrame, w::Interval{Int}, ::_IA_A) = zip(Iterators.repeated(w.y), w.y+1:X(fr)+1)

This pattern is generally convenient; it can, however, be bypassed, although this requires defining two additional methods in order to resolve dispatch ambiguities. When defining a new frame type FR{W}, one can resolve the ambiguities and define a custom accessibles method by providing these three methods:

# access worlds through relation `r`
function accessibles(
    fr::FR{W},
    w::W,
    r::AbstractRelation,
) where {W<:AbstractWorld}
    ...
end

# access current world
function accessibles(
    fr::FR{W},
    w::W,
    r::IdentityRel,
) where {W<:AbstractWorld}
    [w]
end

# access all worlds
function accessibles(
    fr::FR{W},
    w::W,
    r::GlobalRel,
) where {W<:AbstractWorld}
    allworlds(fr)
end

In general, it should be true that collect(accessibles(fr, w, r)) isa AbstractWorlds{W}.

See also AbstractWorld, AbstractRelation, AbstractMultiModalFrame.

SoleData.minify — Function

minify(dataset::D1)::Tuple{D2,Function} where {D1,D2}

Return a minified version of a dataset, as well as a backmap for reverting to the original dataset. Dataset minification remaps each scalar values in the dataset to a new value such that the overall order of the values is preserved; the output dataset is smaller in size, since it relies on values of type UInt8, UInt16, UInt32, etc.

See also isminifiable.

Logisets

SoleData.BASE_FEATURE_ALIASES — Constant

Syntaxstring aliases for standard features, such as "min", "max", "avg".

SoleData.AbstractFeature — Type

abstract type AbstractFeature end

Abstract type for features of worlds of [Kripke structures](https://en.wikipedia.org/wiki/Kripkestructure(model_checking).

See also VarFeature, featvaltype, SoleLogics.AbstractWorld.

SoleData.AbstractUnivariateFeature — Type

abstract type AbstractUnivariateFeature <: VarFeature end

A dimensional feature represented by the application of a function to a single variable of a dimensional channel. For example, it can wrap a scalar function computing how much red a Interval2D world, when interpreted on an image, contains.

See also SoleLogics.Interval, SoleLogics.Interval2D, UnivariateFeature, VarFeature, AbstractFeature.

SoleData.ExplicitFeature — Type

struct ExplicitFeature{T} <: AbstractFeature
    name::String
    featstruct
end

A feature encoded explicitly, for example, as a slice of DimensionalDatasets.UniformFullDimensionalLogiset's feature structure.

See also AbstractFeature.

SoleData.Feature — Type

struct Feature{A} <: AbstractFeature
    atom::A
end

A feature solely identified by an atom (e.g., a string with its name, a tuple of strings, etc.)

See also AbstractFeature.

SoleData.MultivariateFeature — Type

struct MultivariateFeature{U} <: VarFeature
    f::Function
end

A dimensional feature represented by the application of a function to a dimensional channel. For example, it can wrap a scalar function computing how much a Interval2D world, when interpreted on an image, resembles a horse. Note that the image has a number of spatial variables (3, for the case of RGB), and "resembling a horse" may require a computation involving all variables.

See also SoleLogics.Interval, SoleLogics.Interval2D, AbstractUnivariateFeature, VarFeature, AbstractFeature.

SoleData.UnivariateFeature — Type

struct UnivariateFeature{U} <: AbstractUnivariateFeature
    i_variable::Integer
    f::Function
end

A dimensional feature represented by the application of a generic function f to a single variable of a dimensional channel. For example, it can wrap a scalar function computing how much red a Interval2D world, when interpreted on an image, contains.

See also SoleLogics.Interval, SoleLogics.Interval2D, AbstractUnivariateFeature, VarFeature, AbstractFeature.

SoleData.UnivariateMax — Type

struct UnivariateMax <: AbstractUnivariateFeature
    i_variable::Integer
end

Notable univariate feature computing the maximum value for a given variable.

See also SoleLogics.Interval, SoleLogics.Interval2D, AbstractUnivariateFeature, UnivariateMin, VarFeature, AbstractFeature.

SoleData.UnivariateMin — Type

struct UnivariateMin <: AbstractUnivariateFeature
    i_variable::Integer
end

Notable univariate feature computing the minimum value for a given variable.

See also SoleLogics.Interval, SoleLogics.Interval2D, AbstractUnivariateFeature, UnivariateMax, VarFeature, AbstractFeature.

SoleData.UnivariateNamedFeature — Type

struct UnivariateNamedFeature{U} <: AbstractUnivariateFeature
    i_variable::Integer
    name::String
end

A univariate feature solely identified by its name and reference variable.

See also SoleLogics.Interval, SoleLogics.Interval2D, AbstractUnivariateFeature, VarFeature, AbstractFeature.

SoleData.UnivariateSoftMax — Type

struct UnivariateSoftMax{T<:AbstractFloat} <: AbstractUnivariateFeature
    i_variable::Integer
    alpha::T
end

Univariate feature computing a "softened" version of the maximum value for a given variable.

See also SoleLogics.Interval, SoleLogics.Interval2D, AbstractUnivariateFeature, UnivariateMax, VarFeature, AbstractFeature.

SoleData.UnivariateSoftMin — Type

struct UnivariateSoftMin{T<:AbstractFloat} <: AbstractUnivariateFeature
    i_variable::Integer
    alpha::T
end

Univariate feature computing a "softened" version of the minimum value for a given variable.

See also SoleLogics.Interval, SoleLogics.Interval2D, AbstractUnivariateFeature, UnivariateMin, VarFeature, AbstractFeature.

SoleData.UnivariateValue — Type

struct UnivariateValue <: AbstractUnivariateFeature
    i_variable::Integer
end

Simply the value of a scalar variable (propositional case, when the frame has a single world).

See also SoleLogics.Interval, SoleLogics.Interval2D, AbstractUnivariateFeature, UnivariateMax, VarFeature, AbstractFeature.

SoleData.VarFeature — Type

abstract type VarFeature <: AbstractFeature end

Abstract type for feature functions that can be computed on (multi)variate data. Instances of multivariate datasets have values for a number of variables, which can be used to define logical features.

For example, with dimensional data (e.g., multivariate time series, digital images and videos), features can be computed as the minimum value for a given variable on a specific interval/rectangle/cuboid (in general, a `SoleLogics.GeometricalWorld).

As an example of a dimensional feature, consider min[V1], which computes the minimum for variable 1 for a given world. ScalarConditions such as min[V1] >= 10 can be, then, evaluated on worlds.

See also scalarlogiset, featvaltype, computefeature, SoleLogics.Interval.

SoleData.computefeature — Method

computefeature(f::VarFeature, featchannel; kwargs...)

Compute a feature on a featchannel (i.e., world reading) of an instance.

See also VarFeature.

SoleData.computeunivariatefeature — Method

computeunivariatefeature(f::AbstractUnivariateFeature, varchannel; kwargs...)

Compute a feature on a variable channel (i.e., world reading) of an instance.

See also AbstractUnivariateFeature.

SoleData.featvaltype — Method

featvaltype(dataset, f::VarFeature)

Return the type of the values returned by feature f on logiseed dataset.

See also VarFeature.

SoleData.parsefeature — Method

parsefeature(FT::Type{<:AbstractFeature}, expr::String; kwargs...)

Parse a feature of type FT from its syntaxstring representation. Depending on FT, specifying keyword arguments such as featvaltype::Type may be required or recommended.

See also parsecondition.

SoleData.variable_name — Method

variable_name(
    f::AbstractUnivariateFeature;
    variable_names_map::Union{Nothing,AbstractDict,AbstractVector} = nothing,
    variable_name_prefix::Union{Nothing,String} = "V",
)::String

Return the name of the variable targeted by a univariate feature. By default, an variable name is a number prefixed by "V"; however, variable_names_map or variable_name_prefix can be used to customize variable names. The prefix can be customized by specifying variable_name_prefix. Alternatively, a mapping from string to integer (either via a Dictionary or a Vector) can be passed as variable_names_map. Note that only one in variable_names_map and variable_name_prefix should be provided.

See also parsecondition, ScalarCondition, syntaxstring.

SoleData.AbstractCondition — Type

abstract type AbstractCondition{FT<:AbstractFeature} end

Abstract type for representing conditions that can be interpreted and evaluated on worlds of instances of a logical dataset. In logical contexts, these are wrapped into Atoms.

See also Atom, syntaxstring, ScalarMetaCondition, ScalarCondition.

SoleData.AbstractConditionalAlphabet — Type

abstract type AbstractConditionalAlphabet{C<:ScalarCondition} <: SoleLogics.AbstractAlphabet{C} end

Abstract type for alphabets of conditions.

See also ScalarCondition, ScalarMetaCondition.

SoleData.BoundedScalarConditions — Type

struct BoundedScalarConditions{C<:ScalarCondition} <: AbstractConditionalAlphabet{C}
    grouped_featconditions::Vector{Tuple{<:ScalarMetaCondition,Vector}}
end

A finite alphabet of conditions, grouped by (a finite set of) metaconditions.

See also UnboundedScalarConditions, ScalarCondition, ScalarMetaCondition.

SoleData.FunctionalCondition — Type

struct FunctionalCondition{FT<:AbstractFeature} <: AbstractCondition{FT}
    feature::FT
    f::FT
end

A condition which yields a truth value equal to the value of a function.

See also AbstractFeature.

SoleData.ScalarCondition — Type

struct ScalarCondition{U,FT,M<:ScalarMetaCondition{FT}} <: AbstractCondition{FT}
    metacond::M
    a::U
end

A scalar condition comparing a computed feature value (see ScalarMetaCondition) and a threshold value a. It can be evaluated on a world of an instance of a logical dataset.

For example: $min[V1] ≥ 10$, which translates to "Within this world, the minimum of variable 1 is greater or equal than 10." In this case, the feature a UnivariateMin object.

See also AbstractCondition, ScalarMetaCondition.

SoleData.ScalarMetaCondition — Type

struct ScalarMetaCondition{FT<:AbstractFeature,O<:TestOperator} <: AbstractCondition{FT}
    feature::FT
    test_operator::O
end

A metacondition representing a scalar comparison method. Here, the feature is a scalar function that can be computed on a world of an instance of a logical dataset. A test operator is a binary mathematical relation, comparing the computed feature value and an external threshold value (see ScalarCondition). A metacondition can also be used for representing the infinite set of conditions that arise with a free threshold (see UnboundedScalarConditions): ${min[V1] ≥ a, a ∈ ℝ}$.

See also AbstractCondition, ScalarCondition.

SoleData.UnboundedScalarConditions — Type

struct UnboundedScalarConditions{C<:ScalarCondition} <: AbstractConditionalAlphabet{C}
    metaconditions::Vector{<:ScalarMetaCondition}
end

An infinite alphabet of conditions induced from a finite set of metaconditions. For example, if metaconditions = [ScalarMetaCondition(UnivariateMin(1), ≥)], the alphabet represents the (infinite) set: ${min[V1] ≥ a, a ∈ ℝ}$.

See also BoundedScalarConditions, ScalarCondition, ScalarMetaCondition.

SoleData.ValueCondition — Type

struct ValueCondition{FT<:AbstractFeature} <: AbstractCondition{FT}
    feature::FT
end

A condition which yields a truth value equal to the value of a feature.

See also AbstractFeature.

SoleData.parsecondition — Method

parsecondition(C::Type{<:AbstractCondition}, expr::String; kwargs...)

Parse a condition of type C from its syntaxstring representation. Depending on C, specifying keyword arguments such as featuretype::Type{<:AbstractFeature}, and featvaltype::Type may be required or recommended.

See also parsefeature.

SoleData.representatives — Method

representatives(
    fr::AbstractFrame{W},
    S::W,
    ::AbstractRelation,
    ::AbstractCondition
) where {W<:AbstractWorld}

Return an iterator to the (few) representative accessible worlds that are necessary for computing and propagating truth values through existential modal connectives. When this optimization is possible (e.g., when checking specific formulas on scalar conditions), it allows to further boost "one-step" optimizations (see AbstractOneStepMemoset).

For example, consider a Kripke structure with a 1-dimensional FullDimensionalFrame of length 100, and the problem of checking a formula "⟨L⟩(max[V1] ≥ 10)" on a SoleLogics.Interval SoleLogics.Interval{Int64}(1, 2) (with L being Allen's "Later" relation, see SoleLogics.IA_L). Comparing 10 with the (maximum) "max[V1]" computed on all worlds is the naïve strategy to check the formula. However, in this case, comparing 10 to the "max[V1]" computed on the single Interval SoleLogics.Interval{Int64}(2, 101) suffice to establish whether the structure satisfies the formula. Similar cases arise depending on the relation, feature and test operator (or, better, its aggregator).

Note that this method fallsback to accessibles.

See also SoleLogics.accessibles, ScalarCondition, SoleLogics.AbstractFrame.

SoleData.MultiFormula — Type

struct MultiFormula{F<:Formula} <: AbstractSyntaxStructure
    modforms::Dict{Int,F}
end

A symbolic antecedent that can be checked on a MultiLogiset, associating antecedents to modalities.

SoleData.MultiLogiset — Type

struct MultiLogiset{L<:AbstractModalLogiset}
    modalities  :: Vector{L}
end

A logical dataset composed of different modalities); this structure is useful for representing multimodal datasets in logical terms.

See also AbstractModalLogiset, minify.

SoleData.SupportedLogiset — Type

A logiset, associated to a number of cascading full or one-step memoization structures, that are used for optimizing the checking of formulas.

See also SupportedLogiset, AbstractFullMemoset, AbstractOneStepMemoset, AbstractModalLogiset.

SoleData.AbstractFullMemoset — Type

Abstract type for full memoization structures for checking generic formulas.

These structures can be stacked and coupled with one-step memoization structures (see SupportedLogiset).

See AbstractOneStepMemoset, FullMemoset.

SoleData.AbstractMemoset — Type

abstract type AbstractMemoset{
    W<:AbstractWorld,
    U,
    FT<:AbstractFeature,
    FR<:AbstractFrame,
} <: AbstractModalLogiset{W,U,FT,FR} end

Abstract type for memoization structures to be used when checking formulas on logisets.

See also FullMemoset, SupportedLogiset, AbstractModalLogiset.

SoleData.AbstractOneStepMemoset — Type

Abstract type for one-step memoization structures for checking formulas of type ⟨R⟩p; with these formulas, so-called "one-step" optimizations can be performed.

These structures can be stacked and coupled with full memoization structures (see SupportedLogiset).

See ScalarOneStepMemoset, AbstractFullMemoset, representatives.

SoleData.FullMemoset — Type

A generic, full memoization structure that works for any crisp logic; For each instance of a dataset, this structure associates formulas to the set of worlds where the formula holds; it was introduced by Emerson-Clarke for the well-known model checking algorithm for CTL*.

See also SupportedLogiset, AbstractMemoset, AbstractModalLogiset.

SoleData.ScalarChainedMemoset — Type

A full memoization structure used for checking formulas of scalar conditions on datasets with scalar features. This structure is the equivalent to FullMemoset, but with scalar features some important optimizations can be done.

See also FullMemoset, SupportedLogiset, AbstractModalLogiset.

SoleData.AbstractModalLogiset — Type

abstract type AbstractModalLogiset{
    W<:AbstractWorld,
    U,
    FT<:AbstractFeature,
    FR<:AbstractFrame{W},
} <: AbstractLogiset{SoleLogics.AbstractKripkeStructure} end

Abstract type for logisets, that is, logical datasets for symbolic learning where each instance is a Kripke structure associating feature values to each world. Conditions (see AbstractCondition), and logical formulas with conditional letters can be checked on worlds of instances of the dataset.

See also AbstractCondition, AbstractFeature, SoleLogics.AbstractKripkeStructure, SoleLogics.AbstractInterpretationSet.

SoleData.AbstractScalarOneStepRelationalMemoset — Type

Abstract type for one-step memoization structures for checking formulas of type ⟨R⟩ (f ⋈ t), for a generic relation R that is not the global relation (SoleLogics.globalrel). We refer to these structures as relational memosets.

SoleData.ScalarOneStepMemoset — Type

One-step memoization structures for optimized check of formulas of type ⟨R⟩p, where p wraps a scalar condition, such as MyFeature ≥ 10. With such formulas, scalar one-step optimization can be performed.

For example, checking ⟨R⟩(MyFeature ≥ 10) on a world w of a Kripke structure involves comparing the maximum MyFeature across ws accessible worlds with 10; but the same maximum value can be reused to check sibling formulas such as ⟨R⟩(MyFeature ≥ 100). This sparks the idea of storing and reusing scalar aggregations (e.g., minimum/maximum) over the feature values. Each value refers to a specific world, and an object of type ⟨R⟩(f ⋈ ?), called a "scalar metacondition".

Similar cases arise depending on the relation and the test operator (or, better, its aggregator), and further optimizations can be applied for specific feature types (see representatives).

An immediate special case, however, arises when R is the global relation G since, in such case, a single aggregate value is enough for all worlds within the Kripke structure. Therefore, we differentiate between generic, relational memosets (see AbstractScalarOneStepRelationalMemoset), and global memosets (see AbstractScalarOneStepGlobalMemoset), which are usually much smaller.

Given a logiset X, a ScalarOneStepMemoset covers a set of relations and metaconditions, and it holds both a relational and a global memoset. It can be instantiated via:

ScalarOneStepMemoset(
    X                       :: AbstractModalLogiset{W,U},
    metaconditions          :: AbstractVector{<:ScalarMetaCondition},
    relations               :: AbstractVector{<:AbstractRelation};
    precompute_globmemoset  :: Bool = true,
    precompute_relmemoset   :: Bool = false,
    print_progress          :: Bool = false,
)

If precompute_relmemoset is false, then the relational memoset is simply initialized as an empty structure, and memoization is performed on it upon checking formulas. precompute_globmemoset works similarly.

See SupportedLogiset, ScalarMetaCondition, AbstractOneStepMemoset.

Scalar Logisets

Scalar Dimensional Logisets

Multimodal Logisets